pyproject.toml
[tool.poetry]
name = "ioc-enchmnt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ioc_enchmnt"}]

[tool.poetry.dependencies]
python = "^3.11"
whois = "^0.9.27"
shodan = "^1.29.1"
python-dotenv = "^1.0.0"
vt-py = "^0.17.5"
geopy = "^2.3.0"
celery = "^5.3.1"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
